Convert TileMap(Layer) to a module
Converts TileMap(Layer) to a module that can be excluded from builds with 'module_tilemap_enabled=no'.
