+++ editor/scene/2d/SCsub
@@ -6,4 +6,3 @@ Import("env")
 env.add_source_files(env.editor_sources, "*.cpp")
 
 SConscript("physics/SCsub")
-SConscript("tiles/SCsub")
