+++ modules/tilemap/SCsub
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+from misc.utility.scons_hints import *
+
+Import("env")
+Import("env_modules")
+
+env_tilemap = env_modules.Clone()
+
+# Godot source files
+env_tilemap.add_source_files(env.modules_sources, "*.cpp")
+if env.editor_build:
+    env_tilemap.add_source_files(env.modules_sources, "editor/*.cpp")
