+++ modules/tilemap/config.py
@@ -1,0 +1,23 @@
+def can_build(env, platform):
+    return True
+
+
+def configure(env):
+    pass
+
+
+def get_doc_classes():
+    return [
+        "TileMap",
+        "TileMapLayer",
+        "TileData",
+        "TileMapPattern",
+        "TileSet",
+        "TileSetAtlasSource",
+        "TileSetScenesCollectionSource",
+        "TileSetSource",
+    ]
+
+
+def get_doc_path():
+    return "doc_classes"
