+++ modules/tilemap/doc_classes/TileSetScenesCollectionSource.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<class name="TileSetScenesCollectionSource" inherits="TileSetSource" api_type="core" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../class.xsd">
+<class name="TileSetScenesCollectionSource" inherits="TileSetSource" api_type="core" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../doc/class.xsd">
 	<brief_description>
 		Exposes a set of scenes as tiles for a [TileSet] resource.
 	</brief_description>
