+++ modules/tilemap/editor/tile_atlas_view.h
@@ -30,13 +30,14 @@
 
 #pragma once
 
+#include "../tile_set.h"
+
 #include "editor/gui/editor_zoom_widget.h"
 #include "scene/gui/box_container.h"
 #include "scene/gui/button.h"
 #include "scene/gui/center_container.h"
 #include "scene/gui/label.h"
 #include "scene/gui/margin_container.h"
-#include "scene/resources/2d/tile_set.h"
 
 class ViewPanner;
