+++ modules/tilemap/editor/tile_map_layer_editor.cpp
@@ -30,6 +30,10 @@
 
 #include "tile_map_layer_editor.h"
 
+#include "../tile_map.h"
+#include "../tile_map_layer.h"
+#include "tiles_editor_plugin.h"
+
 #include "core/input/input.h"
 #include "core/math/geometry_2d.h"
 #include "core/math/random_pcg.h"
@@ -40,13 +44,10 @@
 #include "editor/editor_undo_redo_manager.h"
 #include "editor/inspector/editor_resource_preview.h"
 #include "editor/inspector/multi_node_edit.h"
-#include "editor/scene/2d/tiles/tiles_editor_plugin.h"
 #include "editor/scene/canvas_item_editor_plugin.h"
 #include "editor/settings/editor_command_palette.h"
 #include "editor/settings/editor_settings.h"
 #include "editor/themes/editor_scale.h"
-#include "scene/2d/tile_map.h"
-#include "scene/2d/tile_map_layer.h"
 #include "scene/gui/grid_container.h"
 #include "scene/gui/split_container.h"
 #include "scene/main/scene_tree.h"
