+++ modules/tilemap/editor/tile_proxies_manager_dialog.h
@@ -30,10 +30,11 @@
 
 #pragma once
 
+#include "../tile_set.h"
+
 #include "editor/inspector/editor_properties.h"
 #include "scene/gui/dialogs.h"
 #include "scene/gui/item_list.h"
-#include "scene/resources/2d/tile_set.h"
 
 class EditorPropertyVector2i;
 class EditorUndoRedoManager;
