+++ modules/tilemap/editor/tile_set_editor.h
@@ -30,13 +30,14 @@
 
 #pragma once
 
+#include "../tile_set.h"
+#include "atlas_merging_dialog.h"
+#include "tile_proxies_manager_dialog.h"
+#include "tile_set_atlas_source_editor.h"
+#include "tile_set_scenes_collection_source_editor.h"
+
 #include "editor/docks/editor_dock.h"
-#include "editor/scene/2d/tiles/atlas_merging_dialog.h"
-#include "editor/scene/2d/tiles/tile_proxies_manager_dialog.h"
-#include "editor/scene/2d/tiles/tile_set_atlas_source_editor.h"
-#include "editor/scene/2d/tiles/tile_set_scenes_collection_source_editor.h"
 #include "scene/gui/tab_bar.h"
-#include "scene/resources/2d/tile_set.h"
 
 class AcceptDialog;
 class SpinBox;
