+++ modules/tilemap/editor/tile_set_scenes_collection_source_editor.cpp
@@ -30,6 +30,8 @@
 
 #include "tile_set_scenes_collection_source_editor.h"
 
+#include "tile_set_editor.h"
+
 #include "core/io/resource_loader.h"
 #include "core/object/callable_mp.h"
 #include "core/object/class_db.h"
@@ -38,7 +40,6 @@
 #include "editor/file_system/editor_file_system.h"
 #include "editor/gui/editor_file_dialog.h"
 #include "editor/inspector/editor_resource_preview.h"
-#include "editor/scene/2d/tiles/tile_set_editor.h"
 #include "editor/settings/editor_settings.h"
 #include "editor/themes/editor_scale.h"
 #include "scene/gui/button.h"
