+++ modules/tilemap/editor/tiles_editor_plugin.cpp
@@ -30,6 +30,11 @@
 
 #include "tiles_editor_plugin.h"
 
+#include "../tile_map.h"
+#include "../tile_map_layer.h"
+#include "../tile_set.h"
+#include "tile_set_editor.h"
+
 #include "core/object/callable_mp.h"
 #include "core/os/mutex.h"
 #include "core/os/os.h"
@@ -38,15 +43,11 @@
 #include "editor/editor_node.h"
 #include "editor/editor_string_names.h"
 #include "editor/inspector/multi_node_edit.h"
-#include "editor/scene/2d/tiles/tile_set_editor.h"
 #include "editor/scene/canvas_item_editor_plugin.h"
 #include "editor/settings/editor_settings.h"
 #include "editor/themes/editor_scale.h"
-#include "scene/2d/tile_map.h"
-#include "scene/2d/tile_map_layer.h"
 #include "scene/gui/control.h"
 #include "scene/main/scene_tree.h"
-#include "scene/resources/2d/tile_set.h"
 #include "scene/resources/image_texture.h"
 #include "servers/rendering/rendering_server.h"
