+++ modules/tilemap/editor/tiles_editor_plugin.h
@@ -30,10 +30,11 @@
 
 #pragma once
 
+#include "tile_map_layer_editor.h"
+#include "tile_set_editor.h"
+
 #include "core/os/semaphore.h"
 #include "editor/plugins/editor_plugin.h"
-#include "editor/scene/2d/tiles/tile_map_layer_editor.h"
-#include "editor/scene/2d/tiles/tile_set_editor.h"
 
 class TilesEditorUtils : public Object {
 	GDCLASS(TilesEditorUtils, Object);
