+++ modules/tilemap/tile_map.h
@@ -30,9 +30,10 @@
 
 #pragma once
 
-#include "scene/2d/tile_map_layer.h"
+#include "tile_map_layer.h"
+#include "tile_set.h"
+
 #include "scene/property_list_helper.h"
-#include "scene/resources/2d/tile_set.h"
 
 class Control;
 #ifndef NAVIGATION_2D_DISABLED
