+++ modules/tilemap/tile_map_layer.cpp
@@ -30,6 +30,8 @@
 
 #include "tile_map_layer.h"
 
+#include "tile_map.h"
+
 #include "core/config/engine.h"
 #include "core/io/marshalls.h"
 #include "core/math/geometry_2d.h"
@@ -37,7 +39,6 @@
 #include "core/object/callable_mp.h"
 #include "core/object/class_db.h"
 #include "core/templates/a_hash_map.h"
-#include "scene/2d/tile_map.h"
 #include "scene/gui/control.h"
 #include "scene/main/scene_tree.h"
 #include "scene/resources/material.h"
